Align client-rpc with client-p2p (structure)
